working on bounding boxes
bounding boxes need to be multiplied by a new matrix. The matrix of the aabb of the model.

+++ src/resources/events/characterControllerEvents/PitchEvent.cpp
@@ -16,7 +16,7 @@ PitchEvent* PitchEvent::DoClone(const void* data) const
 	const InputSystem::MouseData* mouseData = reinterpret_cast<const InputSystem::MouseData*>(data);
 
 	PitchEvent* pitchEvent = new PitchEvent();
-	pitchEvent->SetPitch(mouseData->mCursorY);
+	pitchEvent->SetPitch(static_cast<float>(mouseData->mCursorY));
 
 	return pitchEvent;
 }
